Feat(authentication): add temporary user authentication
Created an Auth class that handles everything related to authentication,
but password aren't being hashed and salted currently, so it's more of
a temporary solution. Proper authentication will be implemented later.
Also migrated from the mysql1 pub package to mysql_client, because
the previous package wasn't updated in a year and there were problems
when making queries with variables.